telegram_bot/requirements.txt
python-telegram-bot==21.10
google-genai>=1.70.0
